requirements.txt
flask==2.3.3
flask-socketio==5.3.6
eventlet==0.33.3
python-dotenv==1.0.0